backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
httpx>=0.27.0
python-multipart>=0.0.9
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
numpy>=1.26.0
